Improved automatic resource collection
Motivation
----------
Incorrectly gathering resources from the bin and obj folders.  Also hard
to customize the behavior.

Modifications
-------------
Exclude the bin and obj folders by default, and provide a property so
exclusions may be extended as needed.

Change to using AutoResourceExtensions items to provide the extensions
to collect.  This may allow the behavior to be extended easily to other
extensions.

Results
-------
Better resource collection system is available for consumers.