Moved UseDebugPort to IPositronUIBuilder extension
Motivation
----------
UseDebugPort doesn't require access to any internal members of
PositronUiBuilder, just methods exposed on IPositronUiBuilder.  As such,
it could be used on any implementation of IPositronUiBuilder for test
fakes.

Modifications
-------------
Moved implementation of UseDebugPort to PositronUiBuilderExtensions, and
removed it from the IPositronUiBuilder interface.

Results
-------
The same functionality as before is provided, but now UseDebugPort can
be called on IPositronUiBuilder fakes without needing to fake the
implementation.