Allow customization of resource filtering for security
Motivation
----------
Current security paradigm blocks all traffic out of Chromium except for
traffic to the internal Positron server.  This is great for security,
but limiting.

Modifications
-------------
Implement IResourceRequestFilter abstraction layer on top of the
OnBeforeResourceLoad method of RequestHandler.

Add a base class, HostResourceRequestFilter, which can be used to
easily implement simple filtering.

Implement the current functionality in
PositronOnlyResourceRequestFilter, which is the registered filter by
default.

Add a test link to the test application.

Results
-------
Current behavior is unchanged, but users may now relax the filtering
restrictions as desired.  API is asynchronous in case external sources
need to be consulted before allowing traffic.