Issue #12: Fix Ctrl-R causes exception and no reload
Motivation
----------
Logging of the address when Ctrl-R is pressed is causing a thread
exception accessing the WPF control, and then crashing the application.

Modifications
-------------
Change to use the IBrowser instance instead of IBrowserControl to
prevent WPF UI thread issues when logging reloads.

Wrap OnKeyEvent logic in a try/catch to help prevent application crashes
with other exceptions that may occur in this handler.

Results
-------
Reload works again.